同步产品代码expense模块截至20190422 12：08

+++ src/main/java/com/hand/hcf/app/expense/input/domain/ExpInputTaxDist.java
@@ -65,7 +65,7 @@ public class ExpInputTaxDist extends Domain {
      */
     @NotNull
     @TableField("transfer_proportion")
-    private Long transferProportion;
+    private BigDecimal transferProportion;
 
     /**
      * 公司
